Best score updating
Added a public variables for the best Score value and the player Name (just to edit them from the Editor).
Added a public variable for the BestScore UI in the canvas to reference it.
Added a method to update the score taking a name and a value as parameter.
Added a condition to update the best score if the current score breaks it.